requirements.txt
alembic==1.18.1
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.12.1
argon2-cffi==25.1.0
argon2-cffi-bindings==25.1.0
asgi-lifespan==2.1.0
async-timeout==5.0.1
asyncpg==0.31.0
bcrypt==5.0.0
black==25.1.0
certifi==2026.1.4
cffi==2.0.0
cfgv==3.5.0
click==8.3.1
coverage==7.13.5
cryptography==46.0.4
distlib==0.4.0
dnspython==2.8.0
email-validator==2.3.0
environs==14.5.0
fastapi==0.128.0
fastapi-cli==0.0.20
fastapi-cloud-cli==0.11.0
fastapi-device-id==0.1.0
fastapi-limiter==0.1.6
fastar==0.8.0
filelock==3.20.3
flake8==7.3.0
freezegun==1.5.5
greenlet==3.3.1
h11==0.16.0
httpcore==1.0.9
httptools==0.7.1
httpx==0.28.1
identify==2.6.16
idna==3.11
iniconfig==2.3.0
isort==6.0.1
itsdangerous==2.2.0
Jinja2==3.1.6
Mako==1.3.10
markdown-it-py==4.0.0
MarkupSafe==3.0.3
marshmallow==4.2.0
mccabe==0.7.0
mdurl==0.1.2
mypy_extensions==1.1.0
nodeenv==1.10.0
orjson==3.11.5
packaging==25.0
passlib==1.7.4
pathspec==1.0.3
platformdirs==4.5.1
pluggy==1.6.0
pre_commit==4.3.0
pycodestyle==2.14.0
pycparser==3.0
pydantic==2.12.5
pydantic-extra-types==2.11.0
pydantic-settings==2.12.0
pydantic_core==2.41.5
pyflakes==3.4.0
Pygments==2.19.2
PyJWT==2.11.0
pytest==9.0.2
pytest-asyncio==1.3.0
pytest-cov==7.1.0
pytest-env==1.6.0
python-dateutil==2.9.0.post0
python-dotenv==1.2.2
python-json-logger==4.0.0
python-multipart==0.0.21
PyYAML==6.0.3
redis==7.1.0
rich==14.2.0
rich-toolkit==0.17.1
rignore==0.7.6
sentry-sdk==2.49.0
shellingham==1.5.4
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.46
starlette==0.50.0
typer==0.21.1
typing-inspection==0.4.2
typing_extensions==4.15.0
ujson==5.11.0
urllib3==2.6.3
uvicorn==0.40.0
uvloop==0.22.1
virtualenv==20.36.1
watchfiles==1.1.1
websockets==16.0
